CustomUserDetails (UserDetails adapter)
Added CustomUserDetails and corresponding service CustomUserDetails.
The goal of this change is to adapt User entities to Spring Security.
CustomUserDetails is a simple class which wraps a User object,
so that it can be used in SpringSecurity.
CustomUserDetailsService is a simple Service whose sole purpose is to
create a CustomUserDetails instance for a given username if it exists.